third_party/WebKit/LayoutTests/http/tests/inspector/security/main-origin-assigned-despite-request-missing-expected.txt

Tests that the Main Origin is assigned even if there is no matching Request.

Page origin: http://127.0.0.1:8000
Detected main origin: http://127.0.0.1:8000
Group: Main Origin
<SPAN class=title >
http://127.0.0.1:8000
</SPAN>
Group: Unknown / Canceled
<SPAN class=title >
https://foo.test
</SPAN>
<SPAN class=title >
https://bar.test
</SPAN>
